choose yeear> created components and reorganized files
