+++ src/CalendarYearTable.js
@@ -1,0 +1,100 @@
+import React from "react";
+import dayjs from "dayjs";
+import { range, createCells } from "./utils";
+
+function CalendarYearTable({ year }) {
+  const monthDays = range(1, 31);
+  const columns = range(1, 14);
+  const css = { cellBorders: "border border-gray-500" };
+  const cells = createCells(year);
+
+  return (
+    <table className="table-fixed w-full border text-xs bg-white">
+      <thead className={`${css.cellBorders}`}>
+        <MonthsRow css={css} />
+      </thead>
+      <tbody>
+        <tr>
+          {columns.map((x, id) => (
+            <td className={`${css.cellBorders} py-6`} key={id}>
+              &nbsp;
+            </td>
+          ))}
+        </tr>
+        {monthDays.map((monthDay) => {
+          return <DaysRow days={cells[monthDay]} key={monthDay} css={css} />;
+        })}
+      </tbody>
+    </table>
+  );
+}
+
+function MonthsRow({ css }) {
+  const months = [
+    "January",
+    "February",
+    "March",
+    "April",
+    "May",
+    "June",
+    "July",
+    "August",
+    "September",
+    "October",
+    "November",
+    "December",
+  ];
+
+  const thClass = `${css.cellBorders} text-gray-700`;
+  return (
+    <tr>
+      <th className={`w-8 ${thClass}`}></th>
+      {months.map((month, id) => (
+        <th key={id} className={thClass}>
+          {month}
+        </th>
+      ))}
+      <th className={`w-8 ${thClass}`}></th>
+    </tr>
+  );
+}
+
+function DaysRow({ days, css }) {
+  const tdClass = `${css.cellBorders} font-semibold text-center text-gray-700`;
+  const monthDay = dayjs(days[0].date).format("D");
+
+  return (
+    <tr>
+      <td className={tdClass}>{monthDay}</td>
+      {days.map((day, id) =>
+        day.hasDate ? (
+          <DayCell
+            weekday={dayjs(day.date).format("dddd")}
+            css={css}
+            key={id}
+          />
+        ) : (
+          <EmptyCell css={css} key={id} />
+        )
+      )}
+      <td className={tdClass}>{monthDay}</td>
+    </tr>
+  );
+}
+
+function EmptyCell({ css }) {
+  return <td className={`${css.cellBorders}`}></td>;
+}
+
+function DayCell({ weekday, css }) {
+  const isSunday =
+    weekday === "Sunday" ? "bg-red-300 text-red-600" : "text-gray-700 ";
+
+  return (
+    <td className={`px-2 ${css.cellBorders} font-semibold ${isSunday}`}>
+      {weekday[0]}
+    </td>
+  );
+}
+
+export default CalendarYearTable;
